requirements.txt
pydantic>=2.6.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
openai>=1.52.0
mcp>=1.2.0
python-dotenv>=1.0.1
duckduckgo-search>=6.3.0
